fixes2

+++ backend/api/views.py
@@ -1,5 +1,4 @@
 from rest_framework import viewsets, permissions, filters
-from rest_framework.permissions import IsAuthenticated, AllowAny
 from django_filters.rest_framework import DjangoFilterBackend
 from rest_framework.decorators import action
 from rest_framework.response import Response
@@ -42,46 +41,43 @@ class RecipeViewSet(viewsets.ModelViewSet):
         url = f"{request.get_host()}/s/{instance.id}"
         return Response(data={"short-link": url})
 
-    @action(detail=True, methods=['post'], url_path='shopping_cart',
+    @action(detail=True, methods=['post', 'delete'], url_path='shopping_cart',
             permission_classes=[permissions.IsAuthenticated])
-    def add_to_cart(self, request, pk=None):
+    def shopping_cart(self, request, pk=None):
         user = request.user
         recipe = self.get_object()
 
-        if ShoppingCart.objects.filter(user=user, recipe=recipe).exists():
-            return Response({'errors': 'Уже в списке покупок'},
-                            status=status.HTTP_400_BAD_REQUEST)
+        if request.method == 'POST':
+            if ShoppingCart.objects.filter(user=user, recipe=recipe).exists():
+                return Response({'errors': 'Уже в списке покупок'},
+                                status=status.HTTP_400_BAD_REQUEST)
+            ShoppingCart.objects.create(user=user, recipe=recipe)
+            serializer = RecipeShortSerializer(recipe,
+                                               context={'request': request})
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
 
-        ShoppingCart.objects.create(user=user, recipe=recipe)
-        serializer = RecipeShortSerializer(recipe,
-                                           context={'request': request})
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
-
-    @action(detail=True, methods=['delete'], url_path='shopping_cart',
-            permission_classes=[permissions.IsAuthenticated])
-    def remove_from_cart(self, request, pk=None):
-        user = request.user
-        recipe = self.get_object()
-
-        cart_item = (ShoppingCart.objects
-                     .filter(user=user, recipe=recipe).first())
-        if not cart_item:
-            return Response({'errors': 'Этого рецепта не было в списке'},
-                            status=status.HTTP_400_BAD_REQUEST)
-
-        cart_item.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        elif request.method == 'DELETE':
+            cart_item = (ShoppingCart.objects
+                         .filter(user=user, recipe=recipe).first())
+            if not cart_item:
+                return Response({'errors': 'Этого рецепта не было в списке'},
+                                status=status.HTTP_400_BAD_REQUEST)
+            cart_item.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
 
     @action(detail=False, methods=['get'],
-            permission_classes=[IsAuthenticated])
+            permission_classes=[permissions.IsAuthenticated])
     def download_shopping_cart(self, request):
         user = request.user
 
         ingredients = (
             Ingredient.objects
             .filter(recipes__in_cart__user=user)
-            .values(name=F('name'), unit=F('measurement_unit'))
-            .annotate(total_amount=Sum('recipeingredient__amount'))
+            .annotate(
+                total_amount=Sum('recipeingredient__amount'),
+                unit=F('measurement_unit')
+            )
+            .values('name', 'unit', 'total_amount')
             .order_by('name')
         )
 
@@ -96,34 +92,31 @@ class RecipeViewSet(viewsets.ModelViewSet):
         response['Content-Disposition'] = f'attachment; filename="{filename}"'
         return response
 
-    @action(detail=True, methods=['post'], url_path='favorite',
-            permission_classes=[IsAuthenticated])
-    def add_to_favorite(self, request, pk=None):
+    @action(detail=True, methods=['post', 'delete'], url_path='favorite',
+            permission_classes=[permissions.IsAuthenticated])
+    def favorite(self, request, pk=None):
         user = request.user
         recipe = self.get_object()
 
-        if Favorite.objects.filter(user=user, recipe=recipe).exists():
-            return Response({'errors': 'Уже в избранном'},
-                            status=status.HTTP_400_BAD_REQUEST)
+        if request.method == 'POST':
+            if Favorite.objects.filter(user=user, recipe=recipe).exists():
+                return Response({'errors': 'Уже в избранном'},
+                                status=status.HTTP_400_BAD_REQUEST)
 
-        Favorite.objects.create(user=user, recipe=recipe)
-        serializer = RecipeShortSerializer(recipe,
-                                           context={'request': request})
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
+            Favorite.objects.create(user=user, recipe=recipe)
+            serializer = RecipeShortSerializer(recipe,
+                                               context={'request': request})
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
 
-    @action(detail=True, methods=['delete'], url_path='favorite',
-            permission_classes=[IsAuthenticated])
-    def remove_from_favorite(self, request, pk=None):
-        user = request.user
-        recipe = self.get_object()
+        if request.method == 'DELETE':
+            favorite = (Favorite.objects
+                        .filter(user=user, recipe=recipe).first())
+            if not favorite:
+                return Response({'errors': 'Этого рецепта нет в избранном'},
+                                status=status.HTTP_400_BAD_REQUEST)
 
-        favorite = Favorite.objects.filter(user=user, recipe=recipe).first()
-        if not favorite:
-            return Response({'errors': 'Этого рецепта нет в избранном'},
-                            status=status.HTTP_400_BAD_REQUEST)
-
-        favorite.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+            favorite.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
 
 
 class IngredientViewSet(viewsets.ReadOnlyModelViewSet):
@@ -135,34 +128,36 @@ class IngredientViewSet(viewsets.ReadOnlyModelViewSet):
 
 
 class UserViewSet(DjoserUserViewSet):
-    permission_classes = [AllowAny]
 
-    @action(detail=False, methods=['put'], url_path='me/avatar',
-            permission_classes=[IsAuthenticated])
-    def set_avatar(self, request):
-        serializer = AvatarUploadSerializer(instance=request.user,
-                                            data=request.data)
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-        return Response({'avatar': request.user.avatar.url},
-                        status=status.HTTP_200_OK)
+    def get_permissions(self):
+        if self.action in ['list', 'retrieve', 'create']:
+            return [permissions.AllowAny()]
+        return [permissions.IsAuthenticated()]
 
-    @action(detail=False, methods=['delete'], url_path='me/avatar',
-            permission_classes=[IsAuthenticated])
-    def delete_avatar(self, request):
-        user = request.user
-        user.avatar.delete(save=True)
-        return Response(status=status.HTTP_204_NO_CONTENT)
+    @action(detail=False, methods=['put', 'delete'], url_path='me/avatar',
+            permission_classes=[permissions.IsAuthenticated])
+    def avatar(self, request):
+        if request.method == 'PUT':
+            serializer = AvatarUploadSerializer(instance=request.user,
+                                                data=request.data)
+            serializer.is_valid(raise_exception=True)
+            serializer.save()
+            return Response({'avatar': request.user.avatar.url},
+                            status=status.HTTP_200_OK)
+
+        elif request.method == 'DELETE':
+            request.user.avatar.delete(save=True)
+            return Response(status=status.HTTP_204_NO_CONTENT)
 
     @action(detail=False, methods=['get'],
-            permission_classes=[IsAuthenticated])
+            permission_classes=[permissions.IsAuthenticated])
     def me(self, request):
         serializer = self.get_serializer(request.user,
                                          context={'request': request})
         return Response(serializer.data)
 
     @action(detail=False, methods=['post'], url_path='set_password',
-            permission_classes=[IsAuthenticated])
+            permission_classes=[permissions.IsAuthenticated])
     def set_password(self, request):
         serializer = SetPasswordSerializer(data=request.data,
                                            context={'request': request})
@@ -173,8 +168,8 @@ class UserViewSet(DjoserUserViewSet):
 
         return Response(status=status.HTTP_204_NO_CONTENT)
 
-    @action(detail=True, methods=['post'], url_path='subscribe',
-            permission_classes=[IsAuthenticated])
+    @action(detail=True, methods=['post', 'delete'], url_path='subscribe',
+            permission_classes=[permissions.IsAuthenticated])
     def subscribe(self, request, id=None):
         user = request.user
         author = self.get_object()
@@ -183,34 +178,30 @@ class UserViewSet(DjoserUserViewSet):
             return Response({'errors': 'Нельзя подписаться на самого себя.'},
                             status=status.HTTP_400_BAD_REQUEST)
 
-        if Subscription.objects.filter(user=user, author=author).exists():
-            return Response(
-                {'errors': 'Вы уже подписаны на этого пользователя.'},
-                status=status.HTTP_400_BAD_REQUEST)
+        if request.method == 'POST':
+            if Subscription.objects.filter(user=user, author=author).exists():
+                return Response(
+                    {'errors': 'Вы уже подписаны на этого пользователя.'},
+                    status=status.HTTP_400_BAD_REQUEST)
 
-        Subscription.objects.create(user=user, author=author)
-        serializer = SubscriptionSerializer(author,
-                                            context={'request': request})
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
+            Subscription.objects.create(user=user, author=author)
+            serializer = SubscriptionSerializer(author,
+                                                context={'request': request})
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
 
-    @action(detail=True, methods=['delete'], url_path='subscribe',
-            permission_classes=[IsAuthenticated])
-    def unsubscribe(self, request, id=None):
-        user = request.user
-        author = self.get_object()
+        elif request.method == 'DELETE':
+            subscription = Subscription.objects.filter(user=user,
+                                                       author=author).first()
+            if not subscription:
+                return Response(
+                    {'errors': 'Вы не подписаны на этого пользователя.'},
+                    status=status.HTTP_400_BAD_REQUEST)
 
-        subscription = (Subscription.objects
-                        .filter(user=user, author=author).first())
-        if not subscription:
-            return Response(
-                {'errors': 'Вы не подписаны на этого пользователя.'},
-                status=status.HTTP_400_BAD_REQUEST)
-
-        subscription.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+            subscription.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
 
     @action(detail=False, methods=['get'],
-            permission_classes=[IsAuthenticated],
+            permission_classes=[permissions.IsAuthenticated],
             url_path='subscriptions')
     def subscriptions(self, request):
         queryset = User.objects.filter(subscribers__user=request.user)
